requirements.txt
flask
flask-login
flask-wtf
requests
python-dotenv
pyyaml
openai
python-dateutil